Move pylxd.profiles.LXDProfile to pylxd.deprecated.profiles
